+++ be1-go/internal/popserver/message/message_test.go
@@ -8,7 +8,7 @@ import (
 	"io"
 	"os"
 	"popstellar/internal/popserver/generator"
-	"popstellar/internal/popserver/utils"
+	"popstellar/internal/popserver/util"
 	"popstellar/network/socket"
 	"popstellar/validation"
 	"testing"
@@ -23,7 +23,7 @@ func TestMain(m *testing.M) {
 		os.Exit(1)
 	}
 
-	utils.InitUtils(&noLog, schemaValidator)
+	util.InitUtils(&noLog, schemaValidator)
 
 	exitVal := m.Run()
